Create TrimSpace
This method is used to trim spaces at the head and the end of a string. For example, if the raw string is "   Hello World   ",the trimmed string will be "Hello World", you'll see the spaces at the head and the end deleted (doesn't include spaces at other positions) when the method is invoked.